+++ engelsystem-base/src/main/kotlin/info/metadude/kotlin/library/engelsystem/EngelsystemApi.kt
@@ -1,0 +1,12 @@
+package info.metadude.kotlin.library.engelsystem
+
+import okhttp3.OkHttpClient
+
+interface EngelsystemApi {
+
+    fun provideEngelsystemService(
+        baseUrl: String,
+        okHttpClient: OkHttpClient
+    ): EngelsystemService
+
+}